refactor : directory 추가
- Game, UI, generator directory 추가
