+++ src/test/java/racingCar/CarTest.java
@@ -5,6 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingCar.game.Car;
+import racingCar.game.generator.IntGenerator;
+import racingCar.game.generator.RandomGenerator;
 
 public class CarTest {
