+++ src/test/java/racingCar/RacingCarGameTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatNoException;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingCar.game.RacingCarGame;
 
 public class RacingCarGameTest {
